Backend/requirement.txt
fastapi
"uvicorn[standard]"
python-multipart
Pillow
transformers
torch